Analyze Cell Overlaps Between Nodes And Generate Heatmap

Computes and visualizes the percentage overlap between nodes in a `FlowObject`, producing a heatmap of overlap percentages. Optionally, it can also display bar plots of specified parameters such as Enrichment, Entropy, or Average Proportion for the selected nodes.

Usage

AnalyzeNodeOverlaps(
  x,
  select_nodes = FALSE,
  n = NULL,
  graphics = TRUE,
  margin = 8,
  parameter = NULL
)

Arguments

x

A `FlowObject` after PruneGatingTree.

select_nodes

Logical value indicating whether to select nodes interactively. Default is `FALSE`, which selects the top `n` nodes based on significance.

n

Integer specifying the number of top nodes to select if `select_nodes` is `FALSE`. Default is `25`.

graphics

Logical indicating whether to use graphical selection when `select_nodes` is `TRUE`. Default is `TRUE`.

margin

Numeric value specifying the margins for the heatmap plot. Default is `8`.

parameter

Character string specifying the parameter to plot alongside the heatmap. Can be `"Enrichment"`, `"Entropy"`, or `"Average Proportion"`. Default is `NULL`, which does not plot additional parameters.

Value

A matrix containing the percentage overlap between the selected nodes.

See also

Examples

if (FALSE) { # \dontrun{
# Basic usage with default settings
AnalyzeNodeOverlaps(x)

# Select nodes interactively
AnalyzeNodeOverlaps(x, select_nodes = TRUE)
} # }