Add Prune Markers to Nodes

This function recursively adds prune markers to nodes of a tree based on specific conditions related to enrichment, entropy, and average proportion.

Usage

add_prune(node, min_average_proportion = 0.001, theta = 0.02)

Arguments

node

The node to process.

min_average_proportion

The minimum average proportion required to not prune a node.

theta

The threshold added to the current enrichment for comparison.

Value

The node with prune markers added.

See also

Examples

if (FALSE) { # \dontrun{
  node <- add_prune(node, 0.001, 0.02)
} # }