Convert Gating Tree to Graphical Representation

This function converts a gating tree from a `FlowObject` into a graph object which can then be visualized or exported. It supports different modes of operation, allowing users to work with either all nodes, pruned nodes, or extracted top nodes based on previous analyses.

Usage

convertGatingTreeToGraph(
  x,
  mode = "Pruned",
  size_factor = 1,
  fontsize = 20,
  all_labels = TRUE
)

Arguments

x

A `FlowObject` containing the gating tree data.

mode

A character string specifying which part of the gating tree to use: - 'All': Uses all nodes from the gating tree. - 'Pruned': Uses nodes pruned by `PruneGatingTree`. - 'Extracted': Uses top nodes extracted by `ExtractTopNodes`. The default is 'Pruned'.

size_factor

A numeric value used to adjust the size of the nodes in the graph. Default is 1.

fontsize

An integer specifying the font size used for node labels in the graph. Default is 20.

all_labels

A logical indicating whether to label all nodes or not. Default is TRUE.

Value

A graph object that can be further processed with graph rendering or exporting functions.

See also

Examples

if (FALSE) { # \dontrun{
  data(FlowObjectExample)
  graph <- convertGatingTreeToGraph(FlowObjectExample)
  render_graph(graph, width = 600, height = 600)
  export_graph(graph, file_name= "PrunedGatingTree.pdf")
} # }