Create a Child Node in a Gating Tree

Constructs a child node for a gating tree based on the specified gating marker and its state, along with related gating statistics and history.

Usage

createChildNode(
  marker,
  current_marker_state,
  indices,
  available_markers,
  current_entropy,
  current_enrichment,
  average_proportion,
  entropy_scores,
  enrichment_scores,
  history,
  isPositive = TRUE,
  depth,
  usedmarkers,
  path
)

Arguments

marker

Character string of the marker name.

current_marker_state

Current state of all markers at the node.

indices

Indices of the data used in this node.

available_markers

Markers available for further gating.

current_entropy

Current entropy score of the node.

current_enrichment

Current enrichment score of the node.

entropy_scores

Entropy scores dataframe.

enrichment_scores

Enrichment scores dataframe.

history

List object containing the history of previous steps.

isPositive

Logical, indicating if the marker state is positive.

depth

Integer, the depth of the node in the tree.

usedmarkers

Vector of markers already used in previous nodes.

path

The path from the root to the current node.

Value

An object of class 'GatingTreeNode'.

See also

Examples

if (FALSE) { # \dontrun{
marker = "CD4"
state = c(1, 0, 2)
indices = 1:100
markers = c("CD4", "CD8")
entropy = 0.5
enrichment = 0.7
scores = data.frame(score=runif(3))
history = list()
depth = 1
usedmarkers = c("CD4")
path = "rootNode"
createChildNode(marker, state, indices, markers, entropy, enrichment, scores, scores, history, TRUE, depth, usedmarkers, path)
} # }