Create a new FlowObject

This function constructs a new FlowObject using various inputs related to flow cytometry analysis, including data, sample definitions, metadata, and data processing configurations.

Usage

flowObject(
  Data = data.frame(),
  sampledef = new("SampleDef"),
  metadata = list(),
  prep = list(),
  Clustering = list(),
  Gating = list(),
  QCdata = list(),
  Transformation = list()
)

Arguments

Data

A data frame containing flow cytometric expression data.

sampledef

An object of class SampleDef, specifying sample definitions.

metadata

A list containing annotation data for the flow cytometry experiment.

prep

A list of character string vectors defining the samples and controls.

Clustering

A list containing cell identities and clustering data.

Gating

A list containing outputs of gating, including GatingTree objects.

QCdata

A list containing the quality control information for the flow cytometry data.

Transformation

A list containing settings for data transformation, including the log data settings.

Value

An object of class FlowObject, which encapsulates all provided data and settings for flow cytometry analysis.